+++ contracts/interfaces/IStorageStateCommittee.sol
@@ -12,9 +12,9 @@ interface IStorageStateCommittee {
     struct CandidateInfo {
         address candidateContract;
         uint256 indexMembers;
-        uint256 memberJoinedTime;
-        uint256 rewardPeriod;
-        uint256 claimedTimestamp;
+        uint128 memberJoinedTime;
+        uint128 rewardPeriod;
+        uint128 claimedTimestamp;
     }
 
     function ton() external returns (address);
